requirements.txt
flask
Flask-SQLAlchemy
mysql-python
jinja2
gunicorn
python-daemon